refactor: Incorporated renderer and removed logic from index.html
This is a massive refactor and a wip, moved the logic out of the index page and into a plugin renderer so that it may be overwritten by themes. Also abstrated a lot of the logic so that there is a specific class representing each audit record and a manager class to look after the transition between database records and the front end data view.