Async report feature (#8)
* feature: Added database for storing flash audit data

Intention is to move this report to a background scheduled task and store the data in order to prevent long wait times required when running as a foreground report for organisations with a large amount of activities.
This database is the first step in facilitating this async feature.

* wip: Added tasks.php for scheduling

* refactor: Broke URLs into separate fields

The URLs were removed from the course and activity fields into separate fields for clarity and in order to allow use of the link independently if required.
Additionally class naming was fixed for the scheduled task.

* fix: Added missing return value

Missing return value meant upgrade process was broken, fixed

* fix: Fixed incorrect namespace

* refactor: Slim down methods

Reduce method footprints for readability and clarity of purpose by introducing some new, single responsibility methods.

* refactor: Added table for tracking audit result counts

* fix: Type in table name

* refactor: Moved datetime logic to separate class

* refactor: Moved logic from index.html to table class

* fix: Fixed pagination

Fixed table to display limited results based on the table page size settings

* improvement: Added method to determine most recent audit datetime

* refactor: Incorporated renderer and removed logic from index.html

This is a massive refactor and a wip, moved the logic out of the index page and into a plugin renderer so that it may be overwritten by themes. Also abstrated a lot of the logic so that there is a specific class representing each audit record and a manager class to look after the transition between database records and the front end data view.

* fix: Fixed the header display

* fix: Fixed broken cron job

* docs: updated docs and tidy up code

* fix: Fixed code style, removed locallib (unused)

* refactor: Removed unessential interface

* refactor: removed unnecessary code

* bug: removed require for locallib now removed

* fix: Fixed html5present to display string instead of boolean int

* refactor: Removed unneccesary buffering

Removed buffering of HTML output which was not necessary
Added in line breaks to make HTML output prettier.

* refactor: Removed audits table

The audits table for tracking when audits were run was considered redundant, this logic has been moved to use the (mdl_)task_scheduled table instead to draw on the datetime of the latest audit.

* refactor: Altered database structure

Updated the database structure to better reflect the query, making it more reusable and allowing the use of moodle API calls as opposed to extra functions for formatting the data to place in the database

* fix: removed unneccessary class

* refactor: Remove API calls from functions

It was noted that too much unnecessary abstraction has been introduced, and overcomplex functions introduced. As a result the permissions of the `audit_activity` class have been altered and the getter methods removed, while the moodle API calls have been moved to where they are invoked, rather than abstracted away into classes.

* refactor: Tidy up `audit_table` class

* fix: Added missing recordset close

* refactor: removed unnecessary function

* refactor: Use more appropriate API call

* refactor: Moved datetime logic to renderer

The date and time calculation of the flash apocalypse was overcomplex, simplified and moved logic to where it is actually needed, in the renderer.

* refactor: Moved to single db call

Moved the DB call to build the coursecategorynames outside of the loop, to reduce DB calls, passing the coursecategorynames as an argument instead.

* fix: Fixed display of html in downloaded table

* fix: Added missing table sort

* improvement: Added a check to display informative message if audit has not been run